separate out utils into files
