+++ src/ir/utils/mod.rs
@@ -1,0 +1,11 @@
+mod foreign;
+mod index_store;
+mod info_map;
+mod interned;
+mod traversal;
+
+pub use foreign::Foreign;
+pub use index_store::IndexStore;
+pub use info_map::DenseIndexInfo;
+pub use interned::Interned;
+pub use traversal::Traversal;
